+++ src/components/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './useDiagnostics';
+export * from './useTheme';
